Fixed reading and writing the creditor's bank name (tag PayeeSpecifiedCreditorFinancialInstitution of type CreditorFinancialInstitutionType)
Extended the assertions in _loadZUGFeRDComfortRabatteInvoice to compare the expected name of the creditor's bank with the actual one. This reveals a bug in InvoiceDescriptorReader (it also exists in InvoiceDescriptorWriter). While for "PayerSpecifiedDebtorFinancialInstitution" the name would be correctly read from the "Name" tag, "PayeeSpecifiedCreditorFinancialInstitution" incorrectly uses the "AccountName" (which by the way would be a valid but obviously unimportant tag for PayeePartyCreditorFinancialAccount). Fixed the InvoiceDescriptorReader and also the InvoiceDescriptorWriter (when the creditor's bank name was set, the generated XML was not valid according to the specification!).