Update InvoiceDescriptorHtmlRenderer.cs
The namespace was changes from ZUGFeRD.XXX to s2industries.ZUGFeRD.XXX. resourceName in the render contained the old namespace in a hardcoded way so the renderer didnt work anymore.
resourceName now reads the namespace in a dynamic way so the problem is fixed and won't occur if the namespace is changed in the future.