Add LineTotalBasisAmount property to Tax class
Introduced a new property `LineTotalBasisAmount` to the `Tax` class in the `s2industries.ZUGFeRD` namespace, representing the basis amount of the line total. Updated the `AddApplicableTradeTax` method in the `InvoiceDescriptor` class to include an optional `lineTotalBasisAmount` parameter. Modified the method signature and body to set this new property on the `Tax` object. Updated the `GlobalTests` class to include assertions for the new property and adjusted test data accordingly. Updated the `Tax` class documentation to include a summary for the new property.