Still for #791.
I've run: `git add --renormalize .`

Reason: On some, but not all files, when I try to make changes , Github recognises the whole file as modified.
In the Internet it is recommended to run `git add --renormalize .`  and to create a pull request for the result, after the handling of line endings has changed.
This is what I've done.
There are no actual changes in the files. You can check with Beyond Compare.